impoement feedback

+++ nifty6/sugar.py
@@ -498,8 +498,7 @@ def calculate_position(operator, output):
     else:
         cov = 1e-3*output.val.max()**2
     invcov = ScalingOperator(output.domain, cov).inverse
-    # FIXME!!!
-    d = output + invcov.draw_sample(dtype=np.float64, from_inverse=True)
+    d = output + invcov.draw_sample(dtype=output.dtype, from_inverse=True)
     lh = GaussianEnergy(d, invcov) @ operator
     H = StandardHamiltonian(
         lh, ic_samp=GradientNormController(iteration_limit=200))
